# virta-rest-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Virta Companies and Stations API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/stations": {
            "get": {
                "tags": [
                    "Stations"
                ],
                "summary": "Get a list of stations",
                "operationId": "4720e59d102825141a63cf3349253ea0",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Stations"
                ],
                "summary": "Create a new station",
                "operationId": "f64cd740423f20d1b8fb367b4e6d0ec2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Station Name"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 37.7749
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -122.4194
                                    },
                                    "company_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "153 Floreasca Tower"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Station created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StationResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity - Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": "The validation error messages."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/stations/{station}": {
            "get": {
                "tags": [
                    "Stations"
                ],
                "summary": "Get a single station",
                "operationId": "35511543ba9348efce1f66b13196465f",
                "parameters": [
                    {
                        "name": "station",
                        "in": "path",
                        "description": "The ID of the station",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StationResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Station not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Stations"
                ],
                "summary": "Update an existing station",
                "operationId": "680e0f1079630c957bd5354c075120f6",
                "parameters": [
                    {
                        "name": "station",
                        "in": "path",
                        "description": "The ID of the station",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Station Name"
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 37.7749
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -122.4194
                                    },
                                    "company_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Updated Address"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Station updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StationResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Station not found"
                    },
                    "422": {
                        "description": "Unprocessable Entity - Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": "The validation error messages."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Stations"
                ],
                "summary": "Delete a station",
                "operationId": "8019fd737c87fa47230b9607d5a19515",
                "parameters": [
                    {
                        "name": "station",
                        "in": "path",
                        "description": "The ID of the station",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Station deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Station deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Station not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "StationResource": {}
        }
    }
}